Add concept of 'BindingHandlers'
Added the concept of 'BindingHandlers'.  The basic idea is they wrap up
the logic involved in updating the UI with a new value.  This allows
for client code to customize exactly how the UI is updated.

Some potential uses for custom written binding handlers could include:
  - Animating changes
  - Reusable transforms (instead of the current transform block on
    bind())
  - Formatters (ie, could build a BindingHandler that uses a
    NSNumberFormatter to transform an Observable<Int> into a String)
  - "Debouncing"/delaying when changes are applied to the UI
  - Filtering (ie: "only update the UI when value is ...")
  - More complex logic (ie: re-rendering a UIImage when a value
    changes)

A default BindingHandler (DefaultBindingHandler) is provided that
simply updates

Additionally, refactor BindableProperty.bind(subscribable:transform:)
to use a custom BindingHandler for doing the transform.