Computed: coalesce updates
Computed now coalesces updates.  So, if 3 of its dependencies change
in a single run loop, it'll only recompute itself once.

NOTE: Because it uses subscriptions for tracking dependencies, if it
depends on a Computed, it'll take 2 runloops to for a change to bubble
up.  Will be fixing it so it only takes the regular one run loop.